Unify JS/Native implementations
