+++ modules/framework/src-js-native/AsyncTask.scala
@@ -4,7 +4,7 @@ import sbt.testing.Task
 import sbt.testing.{ EventHandler, Logger }
 import scala.concurrent.Future
 
-private[weaver] trait SNTask extends Task {
+private[weaver] trait AsyncTask extends Task {
   def executeFuture(
       eventHandler: EventHandler,
       loggers: Array[Logger]): Future[Unit]
