+++ modules/framework/src-js-native/DogFoodCompat.scala
@@ -18,7 +18,7 @@ private[weaver] trait DogFoodCompat[F[_]] { self: DogFood[F] =>
       maxParallelism: Int)(tasks: List[sbt.testing.Task]): F[Unit] = {
     tasks.traverse { task =>
       self.framework.unsafeRun.fromFuture {
-        task.asInstanceOf[SNTask].executeFuture(eventHandler, Array(logger))
+        task.asInstanceOf[AsyncTask].executeFuture(eventHandler, Array(logger))
       }
     }.map { _ =>
       Reporter.logRunFinished(Array(logger))(
