--- modules/framework/src-native/TaskCompat.scala
@@ -1,6 +1,0 @@
-package weaver
-package framework
-
-private[weaver] object TaskCompat {
-  val lineSeparator = "\n"
-}
